LayoutTests/media/track/video/video-track-mkv-theora-language-expected.txt


Check video tracks' language attributes.

EVENT(canplaythrough)

** Check in-band kind attributes
EXPECTED (video.videoTracks.length == '2') OK
EXPECTED (video.videoTracks.getTrackById('V0').language == 'zh') OK
EXPECTED (video.videoTracks.getTrackById('V1').language == 'ru') OK

END OF TEST
